get database moved to fireFunc
